6-size.c

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-printf("size of char: %lu byte(s)\n", size of(char));
-printf("size of int: %lu byte(s)\n", size of(int));
-printf("size of long int: %lu byte(s)\n", size of(long int));
-printf("size of long long int: %lu byte(s)\n", size of(long long int));
-printf("size of float: %lu byte(s)\n", size of(float));
+printf("size of char: %lu byte(s)\n", (unsigned long)size of(char));
+printf("size of int: %lu byte(s)\n", (unsigned long)size of(int));
+printf("size of long int: %lu byte(s)\n", (unsigned long)size of(long int));
+printf("size of long long int: %lu byte(s)\n", (unsigned long)size of(long long int));
+printf("size of float: %lu byte(s)\n", (unsigned long)size of(float));
 Return(0);
 }
